Protect ClinicalAPI Streaming for partial messages
